moved selectors out of reducers
