+++ client/modules/defaults.js
@@ -18,8 +18,6 @@ export const categoriesById = {
   17: 'Gifting'
 };
 
-export const DEFAULT_CATEGORY_ID = '16';
-
 export const defaultTransactions = [
   {
     id: 1,
