+++ client/modules/rootReducer.js
@@ -1,20 +1,7 @@
 import { combineReducers } from 'redux';
-import { createSelector } from 'reselect';
-
-import categories, { applyCategoryName, getCategories } from './categories';
+import categories from './categories';
 import transactionGrid from './transactionGrid';
-import transactions, { getOutflowByCategory } from './transactions';
-
-
-/**
- * Selectors
- */
-export const getOutflowByCategoryName = createSelector(
-  getOutflowByCategory,
-  getCategories,
-  (trans, cat) => applyCategoryName(trans, cat)
-);
-
+import transactions from './transactions';
 
 /**
  * Routing to be implemented
