+++ client/modules/transactions.js
@@ -1,6 +1,3 @@
-import { createSelector } from 'reselect';
-import formatAmount from 'utils/formatAmount';
-
 import {
   defaultTransactions
 } from './defaults';
@@ -46,70 +43,6 @@ function addTransactionToState(state, action) {
   }];
   return newState;
 }
-
-function totalTransactions(transactions) {
-  return transactions.reduce((total, item) => total + parseFloat(item.value), 0).toFixed(2);
-}
-
-function summarizeTransactions(transactions) {
-  return transactions.reduce((summary, { categoryId, value }) => {
-    const sum = summary.find(item => item.categoryId === categoryId) ||
-      summary[summary.push({ categoryId, value: 0 }) - 1];
-
-    sum.value += Math.abs(value);
-    return summary;
-  }, []);
-}
-
-/**
- * Selectors
- */
-export const getTransactions = state => state.transactions;
-
-export const getInflowTransactions = createSelector(
-  [getTransactions],
-  transactions => transactions.filter(item => item.value > 0)
-);
-
-export const getOutflowTransactions = createSelector(
-  [getTransactions],
-  transactions => transactions.filter(item => item.value < 0)
-);
-
-export const getBalance = createSelector(
-  [getTransactions],
-  transactions => totalTransactions(transactions)
-);
-
-export const getInflowBalance = createSelector(
-  [getInflowTransactions],
-  transactions => totalTransactions(transactions)
-);
-
-export const getOutflowBalance = createSelector(
-  [getOutflowTransactions],
-  transactions => totalTransactions(transactions)
-);
-
-export const getFormattedBalance = createSelector(
-  [getBalance],
-  amount => formatAmount(amount, false)
-);
-
-export const getFormattedInflowBalance = createSelector(
-  [getInflowBalance],
-  amount => formatAmount(amount, false)
-);
-
-export const getFormattedOutflowBalance = createSelector(
-  [getOutflowBalance],
-  amount => formatAmount(amount, false)
-);
-
-export const getOutflowByCategory = createSelector(
-  [getOutflowTransactions],
-  transactions => summarizeTransactions(transactions)
-);
 
 /**
  * Reducer
