+++ client/selectors/transactions.js
@@ -1,0 +1,75 @@
+import { createSelector } from 'reselect';
+import formatAmount from 'utils/formatAmount';
+import { getCategories } from './categories';
+
+function totalTransactions(transactions) {
+  return transactions.reduce((total, item) => total + parseFloat(item.value), 0).toFixed(2);
+}
+
+function summarizeTransactions(transactions) {
+  return transactions.reduce((summary, { categoryId, value }) => {
+    const sum = summary.find(item => item.categoryId === categoryId) ||
+      summary[summary.push({ categoryId, value: 0 }) - 1];
+
+    sum.value += Math.abs(value);
+    return summary;
+  }, []);
+}
+
+const applyCategoryName = (transactions, categories) => transactions.map((transaction) => {
+  transaction.category = categories[transaction.categoryId];
+  return transaction;
+});
+
+export const getTransactions = state => state.transactions;
+
+const getInflowTransactions = createSelector(
+  [getTransactions],
+  transactions => transactions.filter(item => item.value > 0)
+);
+
+const getOutflowTransactions = createSelector(
+  [getTransactions],
+  transactions => transactions.filter(item => item.value < 0)
+);
+
+const getBalance = createSelector(
+  [getTransactions],
+  transactions => totalTransactions(transactions)
+);
+
+const getInflowBalance = createSelector(
+  [getInflowTransactions],
+  transactions => totalTransactions(transactions)
+);
+
+const getOutflowBalance = createSelector(
+  [getOutflowTransactions],
+  transactions => totalTransactions(transactions)
+);
+
+export const getFormattedBalance = createSelector(
+  [getBalance],
+  amount => formatAmount(amount, false)
+);
+
+export const getFormattedInflowBalance = createSelector(
+  [getInflowBalance],
+  amount => formatAmount(amount, false)
+);
+
+export const getFormattedOutflowBalance = createSelector(
+  [getOutflowBalance],
+  amount => formatAmount(amount, false)
+);
+
+const getOutflowByCategory = createSelector(
+  [getOutflowTransactions],
+  transactions => summarizeTransactions(transactions)
+);
+
+export const getOutflowByCategoryName = createSelector(
+  getOutflowByCategory,
+  getCategories,
+  (trans, cat) => applyCategoryName(trans, cat)
+);
